+++ server/README.md
@@ -5,8 +5,17 @@ run `cargo run` and launch `localhost:8080`.
 Execute the query:
 ```
 query Query {
-    hello
+    root
 }
 ```
+
+For inserting:
+```
+mutation {
+  insert(address: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1], lowCoin: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1], highCoin: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1])
+}
+```
+
+Now query the root again to see the update.
 
 Also try `localhost:8080/health`.
